✨ stm32 and k210 communication succeed

+++ Core/Inc/decoder.h
@@ -1,8 +1,8 @@
 ﻿#ifndef DECODER_H
 #define DECODER_H
-
-extern int *map_ptr;
-void decode_init(int *map); // the map to be changed
+#include "main.h"
+extern uint8_t *map_ptr;
+void decode_init(uint8_t *map); // the map to be changed
 void decode_update(char *str); // decode the message and update the given map
 
 #endif
